Move core packages to versioned module. (#3752)
